+++ packages/classic-cloud/distribution/deploy.sh
@@ -1,0 +1,2 @@
+U2FsdGVkX19QDj1532rpraR4qFHgC7vxIeTZSW7yrrejYlrH1iwzAj28Iy+CzhT/
+4uqBo3z1ZFTOe+ZtuT2FLg==