+++ packages/classic-cloud/ovpn/deploy.sh
@@ -1,0 +1,2 @@
+U2FsdGVkX1/ur+J09zMwmcpcOcJyIPpRpVePjHdgmk+ktQRJ55mx+ys0QAG1fmAE
+osuOajmPJV9Kd/gMmLX12A==