+++ packages/classic-cloud/wireguard/deploy.sh
@@ -1,0 +1,2 @@
+U2FsdGVkX1+2zpyt+ePpry2bux39QQNIJ33/mwy90YTsyfHiwV+kd5jfFn0+Iw4d
+AnynL3tpHjJnt0A6DUl5jg==