+++ packages/core/math/src/math.ts
@@ -1,4 +1,5 @@
-import {TransformRangeOptions} from './type';
+import {TransformRangeOptions} from './type.js';
+export * from './unicode-digits.js';
 
 /**
  * Number.isFinite simple polyfill
@@ -9,6 +10,9 @@ if (typeof Number.isFinite !== 'function') {
 
 /**
  * Check the value is number or can convert to a number, for example string ' 123 ' can be converted to 123
+ *
+ *  @param {unknown} value - the value must check numberic.
+ *  @return {boolean} - is number status.
  */
 export function isNumber(value: unknown): boolean {
   if (typeof value === 'number') {
@@ -25,19 +29,19 @@ export function isNumber(value: unknown): boolean {
  *
  * Example:
  *
- * ```js
+ * ```ts
  * transformToRange(5, {in: [0, 10], out: [0, 100]}); // => 50
  * ```
  *
  * Make percentage of any value
  *
- * ```js
+ * ```ts
  * transformToRange(2000, {in: [0, 5000], out: [0, 100]}); // => 40
  * ```
  *
  * Calculate progress-bar with
  *
- * ```js
+ * ```ts
  * const progressOuterWith = 400; //px
  * const gap = 5; //px (the visual gap between progressBar and component outer).
  * const currentProgress = 30; //%
@@ -190,93 +194,4 @@ export function parseDuration(duration: DurationString, toUnit: DurationUnit | '
     throw new Error(`invalid_init`);
   }
   return (durationNumber * unitConversion[durationUnit]) / (toUnit === 'ms' ? 1 : unitConversion[toUnit]);
-}
-
-export const unicodeDigits = {
-  'en': 0x0030,
-  'ar': 0x0660,
-  'fa': 0x06f0,
-  'nko': 0x07c0,
-  'hi': 0x0966, // devanagari
-  'bn': 0x09e6, // bengali
-  'pa': 0x0a66, // gurmukhi, punjabi
-  'gu': 0x0ae6, // gujarati
-  'or': 0x0b66, // oriya
-  'ta': 0x0be6, // tamil
-  'te': 0x0c66, // telugu
-  'kn': 0x0ce6, // kannada
-
-  'mal': 0x0d66, // malayalam
-  'sinhala_lith': 0x0de6,
-  'thai': 0x0e50,
-  'lao': 0x0ed0,
-  'tibetan': 0x0f20,
-  'myanmar': 0x1040,
-  'myanmar_shan': 0x1090,
-  'khmer': 0x17e0,
-  'mongolian': 0x1810,
-  'limbu': 0x1946,
-  'new_tai_lue': 0x19d0,
-  'tai_tham_hora': 0x1a80,
-  'tai_tham_tham': 0x1a90,
-  'balinese': 0x1b50,
-  'sundanese': 0x1bb0,
-  'lepcha': 0x1c40,
-  'ol_chiki': 0x1c50,
-  'vai': 0xa620,
-  'saurashtra': 0xa8d0,
-  'kayah_li': 0xa900,
-  'javanese': 0xa9d0,
-  'myanmar_tai_laing': 0xa9f0,
-  'cham': 0xaa50,
-  'meetei_mayek': 0xabf0,
-  'fullwidth': 0xff10,
-  'osmanya': 0x104a0,
-  'brahmi': 0x11066,
-  'sora_sompeng': 0x110f0,
-  'chakma': 0x11136,
-  'sharada': 0x111d0,
-  'khudawadi': 0x112f0,
-  'newa': 0x11450,
-  'tirhuta': 0x114d0,
-  'modi': 0x11650,
-  'takri': 0x116c0,
-  'ahom': 0x11730,
-  'warang_citi': 0x118e0,
-  'bhaiksuki': 0x11c50,
-  'mro': 0x16a60,
-  'pahawh_hmong': 0x16b50,
-  'mathematical_bold': 0x1d7ce,
-  'mathematical_double-struck': 0x1d7d8,
-  'mathematical_sans-serif': 0x1d7e2,
-  'mathematical_sans-serif_bold': 0x1d7ec,
-  'mathematical_monospace': 0x1d7f6,
-  'fula': 0x1e950, // adlam script in fula lang
-} as const;
-
-export type LangKeys = keyof typeof unicodeDigits;
-const allLangKeys: Array<LangKeys> = Object.keys(unicodeDigits) as Array<LangKeys>;
-
-/**
- * Replace all digital number in all languages to requested language.
- *
- * Example:
- *
- * ```js
- * translateUnicodeDigits('123 ߁߂߃ ੧੨੩ ٣٤٦', 'fa'); // ۱۲۳ ۱۲۳ ۱۲۳ ۳۴۶
- * ```
- */
-export function translateUnicodeDigits(
-    string: string,
-    toLanguage: LangKeys = 'en',
-    fromLanguages: Array<LangKeys> = allLangKeys,
-): string {
-  const toLangZeroCode = unicodeDigits[toLanguage];
-  for (const langKey of fromLanguages) {
-    const fromLangZeroCode = unicodeDigits[langKey];
-    for (let n = 0; n < 10; n++) {
-      string = string.replaceAll(String.fromCharCode(fromLangZeroCode + n), String.fromCharCode(toLangZeroCode + n));
-    }
-  }
-  return string;
 }