+++ packages/core/router/src/router.ts
@@ -1,22 +1,25 @@
-import {joinParameterList, logger, routeSignalProvider} from './core';
-import {routeChangeSignal} from './signal';
-import {clickTrigger} from './trigger-click';
-import {popstateTrigger} from './trigger-popstate';
+import {joinParameterList, logger, routeSignalProvider} from './core.js';
+import {routeChangeSignal} from './signal.js';
+import {clickTrigger} from './trigger-click.js';
+import {popstateTrigger} from './trigger-popstate.js';
 
-import type {InitOptions, Route, RoutesConfig} from './type';
+import type {InitOptions, Route, RoutesConfig} from './type.js';
 import type {SignalInterface} from '@alwatr/signal';
 
 export {routeChangeSignal};
-export type {Route, RequestRouteParam, RoutesConfig} from './type';
+export type {Route, RequestRouteParam, RoutesConfig} from './type.js';
 
 /**
  * Initial and config the Router.
  */
-function initial(options?: InitOptions): void {
+function initial(options: InitOptions = {}): void {
+  options.clickTrigger ??= true;
+  options.popstateTrigger ??= true;
+
   logger.logMethodArgs('initialRouter', {options});
 
-  clickTrigger.enable = options?.clickTrigger ?? true;
-  popstateTrigger.enable = options?.popstateTrigger ?? true;
+  clickTrigger.enable = options.clickTrigger;
+  popstateTrigger.enable = options.popstateTrigger;
 
   routeChangeSignal.setProvider(routeSignalProvider, {debounce: true, receivePrevious: true});
 
@@ -24,7 +27,7 @@ function initial(options?: InitOptions): void {
   if (!routeChangeSignal.dispatched) {
     const {pathname, search, hash} = window.location;
     // Don't use `routeChangeSignal.request()` because we need set the route value immediately.
-    routeChangeSignal.dispatch(routeSignalProvider({pathname, search, hash, pushState: false}));
+    routeChangeSignal.dispatch(routeSignalProvider({pathname, search, hash, pushState: false}), {debounce: false});
   }
 }