+++ packages/core/storage/src/storage.ts
@@ -5,9 +5,9 @@ import exitHook from 'exit-hook';
 
 import {readJsonFile, writeJsonFile} from './util.js';
 
-import type {DocumentObject, DocumentListStorage, AlwatrStorageConfig} from './type';
+import type {DocumentObject, DataStorage, AlwatrStorageConfig} from './type.js';
 
-export {DocumentObject, DocumentListStorage, AlwatrStorageConfig as Config};
+export {DocumentObject, DataStorage, AlwatrStorageConfig as Config};
 
 alwatrRegisteredList.push({
   name: '@alwatr/storage',
@@ -18,15 +18,56 @@ alwatrRegisteredList.push({
  * Elegant micro in-memory json-like storage with disk backed,
  * Fastest NoSQL Database written in tiny TypeScript ES module.
  *
- * @param {string} name Storage name like database table name.
- * @param {string} pathPrefix Saved file path prefix (default is `data`).
- *
  * Example:
  *
  * ```ts
- * import {AlwatrStorage, DocumentObject} from '@alwatr/storage';
- * interface User extends DocumentObject {...}
- * const db = new AlwatrStorage<User>('user-list');
+ * import {AlwatrStorage} from '@alwatr/storage';
+ *
+ * import type {DocumentObject} from '@alwatr/storage';
+ *
+ * interface User extends DocumentObject {
+ *   fname: string;
+ *   lname: string;
+ *   email: string;
+ *   token?: string;
+ * }
+ *
+ * const db = new AlwatrStorage<User>({
+ *   name: 'user-list',
+ *   path: 'db',
+ *   saveBeautiful: true,
+ *   debug: true,
+ * });
+ *
+ * console.log('db loaded and ready to access.');
+ *
+ * let ali = db.get('alimd');
+ *
+ * if (ali == null) {
+ *   console.log('ali not found');
+ *   ali = {
+ *     _id: 'alimd',
+ *     _updatedBy: 'demo',
+ *     fname: 'Ali',
+ *     lname: 'Mihandoost',
+ *     email: 'ali@mihandoost.com',
+ *   };
+ * }
+ * else {
+ *   console.log('ali found: %o', ali);
+ *   ali.token = Math.random().toString(36).substring(2, 15);
+ * }
+ *
+ * db.set(ali);
+ *
+ * db.set({
+ *   _id: 'fmd',
+ *   _updatedBy: 'demo',
+ *   fname: 'Fatemeh',
+ *   lname: 'Mihandoost',
+ *   email: 'Fatemeh@mihandoost.com',
+ *   token: Math.random().toString(36).substring(2, 15),
+ * });
  * ```
  */
 export class AlwatrStorage<DocumentType extends DocumentObject> {
@@ -56,13 +97,11 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
   hasUnsavedChanges = false;
 
   protected _logger;
-  protected _storage: DocumentListStorage<DocumentType> = {};
+  protected _storage: DataStorage<DocumentType>;
   protected _keys: Array<string> | null = null;
 
   /**
    * All document ids in array.
-   *
-   * Contain `_last`!
    */
   get keys(): Array<string> {
     if (this._keys === null) {
@@ -72,10 +111,14 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
   }
 
   /**
-   * Size of the storage (count `_last`).
+   * Size of the storage.
    */
   get length(): number {
     return this.keys.length;
+  }
+
+  protected get _newStorage(): DataStorage<DocumentType> {
+    return {ok: true, data: {}};
   }
 
   constructor(config: AlwatrStorageConfig) {
@@ -89,15 +132,26 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
     this.saveBeautiful = config.saveBeautiful || false;
 
     exitHook(this.forceSave);
-    this.load();
+    this._storage = this.load();
   }
 
   /**
-   * Load process like open/parse storage file.
+   * load storage file.
    */
-  private load(): void {
+  protected load(): DataStorage<DocumentType> {
     this._logger.logMethodArgs('load', {path: this.storagePath});
-    this._storage = readJsonFile<DocumentListStorage<DocumentType>>(this.storagePath) ?? {};
+
+    const storage = readJsonFile<DataStorage<DocumentType>>(this.storagePath);
+
+    if (storage === null) {
+      return this._newStorage;
+    }
+
+    if (storage.ok !== true) {
+      throw new Error('invalid_data');
+    }
+
+    return storage;
   }
 
   /**
@@ -110,7 +164,7 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
    * ```
    */
   has(documentId: string): boolean {
-    return this._storage[documentId] != null;
+    return this._storage.data[documentId] != null;
   }
 
   /**
@@ -130,9 +184,8 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
   get(documentId: string, fastInstance?: boolean): DocumentType | null {
     this._logger.logMethodArgs('get', documentId);
 
-    const documentObject = this._storage[documentId];
+    const documentObject = this._storage.data[documentId];
     if (typeof documentObject === 'string') {
-      // for example _last
       return this.get(documentObject);
     }
     else if (documentObject == null) {
@@ -166,8 +219,8 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
   set(documentObject: DocumentType, fastInstance?: boolean): DocumentType {
     this._logger.logMethodArgs('set', documentObject._id);
 
-    const oldData = this._storage[documentObject._id];
-    if (oldData == null) this._keys = null; // Clear cached keys on new docId
+    const oldData = this._storage.data[documentObject._id];
+    if (oldData == null) this._keys = null; // Clear cached keys
 
     if (fastInstance !== true) {
       documentObject = JSON.parse(JSON.stringify(documentObject));
@@ -179,8 +232,7 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
     documentObject._createdBy = oldData?._createdBy ?? documentObject._updatedBy;
     documentObject._rev = (oldData?._rev ?? 0) + 1;
 
-    this._storage._last = documentObject._id;
-    this._storage[documentObject._id] = documentObject;
+    this._storage.data[documentObject._id] = documentObject;
 
     this.save();
     return documentObject;
@@ -198,11 +250,15 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
   remove(documentId: string): boolean {
     this._logger.logMethodArgs('remove', documentId);
 
-    if (this._storage[documentId] == null) {
+    if (this._storage.data[documentId] == null) {
       return false;
     }
     // else
-    delete this._storage[documentId];
+    delete this._storage.data[documentId];
+
+    // Clear cached keys
+    this._keys = null;
+
     this.save();
     return true;
   }
@@ -224,7 +280,6 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
   async forAll(callbackfn: (documentObject: DocumentType) => void | false | Promise<void | false>): Promise<void> {
     const keys = this.keys;
     for (const documentId of keys) {
-      if (documentId === '_last') continue; // prevent to duplicate latest key.
       const documentObject = this.get(documentId);
       if (documentObject != null) {
         const retVal = await callbackfn(documentObject);
@@ -264,13 +319,19 @@ export class AlwatrStorage<DocumentType extends DocumentObject> {
   }
 
   /**
-   * Unload storage data and free ram usage.
+   * Unload storage data and free ram usage (auto saved before unload).
+   *
+   * Example:
+   *
+   * ```ts
+   * userStorage.unload();
+   * delete userStorage;
+   * ```
    */
   unload(): void {
     this._logger.logMethod('unload');
-    if (this.hasUnsavedChanges) {
-      this.forceSave();
-    }
-    this._storage = {};
+    this.forceSave();
+    this._keys = null;
+    this._storage = this._newStorage;
   }
 }